semi final tests

+++ db/migrate/20240804153636_drop_todos_table.rb
@@ -1,9 +1,14 @@
-class DropTodosTable < ActiveRecord::Migration[7.1]
+class DropTodosTable < ActiveRecord::Migration[7.0]
   def up
-    drop_table :todos
+    if table_exists?(:todos)
+      drop_table :todos
+    end
   end
 
   def down
-    raise ActiveRecord::IrreversibleMigration
+    create_table :todos do |t|
+      t.string :name
+      t.timestamps
+    end
   end
 end
